+++ sdk/textanalytics/azure-ai-textanalytics/azure/ai/textanalytics/_models.py
@@ -330,24 +330,31 @@ class DetectedLanguage(DictMixin):
     :ivar confidence_score: A confidence score between 0 and 1. Scores close
         to 1 indicate 100% certainty that the identified language is true.
     :vartype confidence_score: float
+    :ivar Optional[str] script: Identifies the script of the input document. Possible values: "Latin".
+
+    .. versionadded:: 2022-10-01-preview
+        The *script* property.
     """
 
     def __init__(self, **kwargs):
         self.name = kwargs.get("name", None)
         self.iso6391_name = kwargs.get("iso6391_name", None)
         self.confidence_score = kwargs.get("confidence_score", None)
+        self.script = kwargs.get("script", None)
 
     @classmethod
     def _from_generated(cls, language):
+        script = language.script if hasattr(language, "script") else None
         return cls(
             name=language.name,
             iso6391_name=language.iso6391_name,
             confidence_score=language.confidence_score,
+            script=script
         )
 
     def __repr__(self):
-        return "DetectedLanguage(name={}, iso6391_name={}, confidence_score={})".format(
-            self.name, self.iso6391_name, self.confidence_score
+        return "DetectedLanguage(name={}, iso6391_name={}, confidence_score={}, script={})".format(
+            self.name, self.iso6391_name, self.confidence_score, self.script
         )[:1024]
 
 
@@ -369,9 +376,15 @@ class RecognizeEntitiesResult(DictMixin):
         field will contain information about the document payload.
     :vartype statistics:
         Optional[~azure.ai.textanalytics.TextDocumentStatistics]
+    :ivar detected_language: If 'language' was set to 'auto' for the document in the request this
+        field will contain a 2 letter ISO 639-1 representation of the language detected for this document.
+    :vartype detected_language: Optional[~azure.ai.textanalytics.DetectedLanguage]
     :ivar bool is_error: Boolean check for error item when iterating over list of
         results. Always False for an instance of a RecognizeEntitiesResult.
     :ivar str kind: The text analysis kind - "EntityRecognition".
+
+    .. versionadded:: 2022-10-01-preview
+        The *detected_language* property.
     """
 
     def __init__(self, **kwargs):
@@ -379,15 +392,18 @@ class RecognizeEntitiesResult(DictMixin):
         self.entities = kwargs.get("entities", None)
         self.warnings = kwargs.get("warnings", [])
         self.statistics = kwargs.get("statistics", None)
+        self.detected_language = kwargs.get("detected_language", None)
         self.is_error: Literal[False] = False
         self.kind: Literal["EntityRecognition"] = "EntityRecognition"
 
     def __repr__(self):
-        return "RecognizeEntitiesResult(id={}, entities={}, warnings={}, statistics={}, is_error={})".format(
+        return "RecognizeEntitiesResult(id={}, entities={}, warnings={}, statistics={}, " \
+               "detected_language={}, is_error={})".format(
             self.id,
             repr(self.entities),
             repr(self.warnings),
             repr(self.statistics),
+            repr(self.detected_language),
             self.is_error,
         )[
             :1024
@@ -414,9 +430,15 @@ class RecognizePiiEntitiesResult(DictMixin):
         field will contain information about the document payload.
     :vartype statistics:
         Optional[~azure.ai.textanalytics.TextDocumentStatistics]
+    :ivar detected_language: If 'language' was set to 'auto' for the document in the request this
+        field will contain a 2 letter ISO 639-1 representation of the language detected for this document.
+    :vartype detected_language: Optional[~azure.ai.textanalytics.DetectedLanguage]
     :ivar bool is_error: Boolean check for error item when iterating over list of
         results. Always False for an instance of a RecognizePiiEntitiesResult.
     :ivar str kind: The text analysis kind - "PiiEntityRecognition".
+
+    .. versionadded:: 2022-10-01-preview
+        The *detected_language* property.
     """
 
     def __init__(self, **kwargs):
@@ -425,18 +447,20 @@ class RecognizePiiEntitiesResult(DictMixin):
         self.redacted_text = kwargs.get("redacted_text", None)
         self.warnings = kwargs.get("warnings", [])
         self.statistics = kwargs.get("statistics", None)
+        self.detected_language = kwargs.get('detected_language', None)
         self.is_error: Literal[False] = False
         self.kind: Literal["PiiEntityRecognition"] = "PiiEntityRecognition"
 
     def __repr__(self):
         return (
             "RecognizePiiEntitiesResult(id={}, entities={}, redacted_text={}, warnings={}, "
-            "statistics={}, is_error={})".format(
+            "statistics={}, detected_language={}, is_error={})".format(
                 self.id,
                 repr(self.entities),
                 self.redacted_text,
                 repr(self.warnings),
                 repr(self.statistics),
+                repr(self.detected_language),
                 self.is_error,
             )[:1024]
         )
@@ -470,12 +494,15 @@ class AnalyzeHealthcareEntitiesResult(DictMixin):
         FHIR compatible object for consumption in other Healthcare tools. For additional
         information see https://www.hl7.org/fhir/overview.html.
     :vartype fhir_bundle: Optional[dict[str, any]]
+    :ivar detected_language: If 'language' was set to 'auto' for the document in the request this
+        field will contain a 2 letter ISO 639-1 representation of the language detected for this document.
+    :vartype detected_language: Optional[~azure.ai.textanalytics.DetectedLanguage]
     :ivar bool is_error: Boolean check for error item when iterating over list of
         results. Always False for an instance of a AnalyzeHealthcareEntitiesResult.
     :ivar str kind: The text analysis kind - "Healthcare".
 
     .. versionadded:: 2022-10-01-preview
-        The *fhir_bundle* property.
+        The *fhir_bundle* and *detected_language* properties.
     """
 
     def __init__(self, **kwargs):
@@ -485,6 +512,7 @@ class AnalyzeHealthcareEntitiesResult(DictMixin):
         self.warnings = kwargs.get("warnings", [])
         self.statistics = kwargs.get("statistics", None)
         self.fhir_bundle = kwargs.get("fhir_bundle", None)
+        self.detected_language = kwargs.get('detected_language', None)
         self.is_error: Literal[False] = False
         self.kind: Literal["Healthcare"] = "Healthcare"
 
@@ -515,18 +543,22 @@ class AnalyzeHealthcareEntitiesResult(DictMixin):
                 healthcare_result.statistics
             ),
             fhir_bundle=fhir_bundle,
+            detected_language=DetectedLanguage._from_generated(  # pylint: disable=protected-access
+                healthcare_result.detected_language
+            ) if hasattr(healthcare_result, "detected_language") and healthcare_result.detected_language else None
         )
 
     def __repr__(self):
         return (
             "AnalyzeHealthcareEntitiesResult(id={}, entities={}, entity_relations={}, warnings={}, "
-            "statistics={}, fhir_bundle={}, is_error={})".format(
+            "statistics={}, fhir_bundle={}, detected_language={}, is_error={})".format(
                 self.id,
                 repr(self.entities),
                 repr(self.entity_relations),
                 repr(self.warnings),
                 repr(self.statistics),
                 self.fhir_bundle,
+                repr(self.detected_language),
                 self.is_error,
             )[:1024]
         )
@@ -1021,9 +1053,15 @@ class ExtractKeyPhrasesResult(DictMixin):
         field will contain information about the document payload.
     :vartype statistics:
         Optional[~azure.ai.textanalytics.TextDocumentStatistics]
+    :ivar detected_language: If 'language' was set to 'auto' for the document in the request this
+        field will contain a 2 letter ISO 639-1 representation of the language detected for this document.
+    :vartype detected_language: Optional[~azure.ai.textanalytics.DetectedLanguage]
     :ivar bool is_error: Boolean check for error item when iterating over list of
         results. Always False for an instance of a ExtractKeyPhrasesResult.
     :ivar str kind: The text analysis kind - "KeyPhraseExtraction".
+
+    .. versionadded:: 2022-10-01-preview
+        The *detected_language* property.
     """
 
     def __init__(self, **kwargs):
@@ -1031,15 +1069,18 @@ class ExtractKeyPhrasesResult(DictMixin):
         self.key_phrases = kwargs.get("key_phrases", None)
         self.warnings = kwargs.get("warnings", [])
         self.statistics = kwargs.get("statistics", None)
+        self.detected_language = kwargs.get('detected_language', None)
         self.is_error: Literal[False] = False
         self.kind: Literal["KeyPhraseExtraction"] = "KeyPhraseExtraction"
 
     def __repr__(self):
-        return "ExtractKeyPhrasesResult(id={}, key_phrases={}, warnings={}, statistics={}, is_error={})".format(
+        return "ExtractKeyPhrasesResult(id={}, key_phrases={}, warnings={}, statistics={}, " \
+               "detected_language={}, is_error={})".format(
             self.id,
             self.key_phrases,
             repr(self.warnings),
             repr(self.statistics),
+            repr(self.detected_language),
             self.is_error,
         )[
             :1024
@@ -1064,9 +1105,15 @@ class RecognizeLinkedEntitiesResult(DictMixin):
         field will contain information about the document payload.
     :vartype statistics:
         Optional[~azure.ai.textanalytics.TextDocumentStatistics]
+    :ivar detected_language: If 'language' was set to 'auto' for the document in the request this
+        field will contain a 2 letter ISO 639-1 representation of the language detected for this document.
+    :vartype detected_language: Optional[~azure.ai.textanalytics.DetectedLanguage]
     :ivar bool is_error: Boolean check for error item when iterating over list of
         results. Always False for an instance of a RecognizeLinkedEntitiesResult.
     :ivar str kind: The text analysis kind - "EntityLinking".
+
+    .. versionadded:: 2022-10-01-preview
+        The *detected_language* property.
     """
 
     def __init__(self, **kwargs):
@@ -1074,15 +1121,18 @@ class RecognizeLinkedEntitiesResult(DictMixin):
         self.entities = kwargs.get("entities", None)
         self.warnings = kwargs.get("warnings", [])
         self.statistics = kwargs.get("statistics", None)
+        self.detected_language = kwargs.get('detected_language', None)
         self.is_error: Literal[False] = False
         self.kind: Literal["EntityLinking"] = "EntityLinking"
 
     def __repr__(self):
-        return "RecognizeLinkedEntitiesResult(id={}, entities={}, warnings={}, statistics={}, is_error={})".format(
+        return "RecognizeLinkedEntitiesResult(id={}, entities={}, warnings={}, statistics={}, " \
+               "detected_language={}, is_error={})".format(
             self.id,
             repr(self.entities),
             repr(self.warnings),
             repr(self.statistics),
+            repr(self.detected_language),
             self.is_error,
         )[
             :1024
@@ -1116,9 +1166,15 @@ class AnalyzeSentimentResult(DictMixin):
     :ivar sentences: Sentence level sentiment analysis.
     :vartype sentences:
         list[~azure.ai.textanalytics.SentenceSentiment]
+    :ivar detected_language: If 'language' was set to 'auto' for the document in the request this
+        field will contain a 2 letter ISO 639-1 representation of the language detected for this document.
+    :vartype detected_language: Optional[~azure.ai.textanalytics.DetectedLanguage]
     :ivar bool is_error: Boolean check for error item when iterating over list of
         results. Always False for an instance of a AnalyzeSentimentResult.
     :ivar str kind: The text analysis kind - "SentimentAnalysis".
+
+    .. versionadded:: 2022-10-01-preview
+        The *detected_language* property.
     """
 
     def __init__(self, **kwargs):
@@ -1128,19 +1184,21 @@ class AnalyzeSentimentResult(DictMixin):
         self.statistics = kwargs.get("statistics", None)
         self.confidence_scores = kwargs.get("confidence_scores", None)
         self.sentences = kwargs.get("sentences", None)
+        self.detected_language = kwargs.get("detected_language", None)
         self.is_error: Literal[False] = False
         self.kind: Literal["SentimentAnalysis"] = "SentimentAnalysis"
 
     def __repr__(self):
         return (
             "AnalyzeSentimentResult(id={}, sentiment={}, warnings={}, statistics={}, confidence_scores={}, "
-            "sentences={}, is_error={})".format(
+            "sentences={}, detected_language={}, is_error={})".format(
                 self.id,
                 self.sentiment,
                 repr(self.warnings),
                 repr(self.statistics),
                 repr(self.confidence_scores),
                 repr(self.sentences),
+                repr(self.detected_language),
                 self.is_error,
             )[:1024]
         )
@@ -1401,14 +1459,18 @@ class TextDocumentInput(DictMixin, MultiLanguageInput):
     :keyword str id: Required. Unique, non-empty document identifier.
     :keyword str text: Required. The input text to process.
     :keyword str language: This is the 2 letter ISO 639-1 representation
-     of a language. For example, use "en" for English; "es" for Spanish etc. If
+     of a language. For example, use "en" for English; "es" for Spanish etc.
+     For automatic language detection, use "auto" (Only supported by long-running
+     operation APIs with API version 2022-10-01-preview or newer). If
      not set, uses "en" for English as default.
     :ivar id: Required. Unique, non-empty document identifier.
     :vartype id: str
     :ivar text: Required. The input text to process.
     :vartype text: str
     :ivar language: This is the 2 letter ISO 639-1 representation
-     of a language. For example, use "en" for English; "es" for Spanish etc. If
+     of a language. For example, use "en" for English; "es" for Spanish etc.
+     For automatic language detection, use "auto" (Only supported by long-running
+     operation APIs with API version 2022-10-01-preview or newer). If
      not set, uses "en" for English as default.
     :vartype language: Optional[str]
     """
@@ -2244,9 +2306,15 @@ class RecognizeCustomEntitiesResult(DictMixin):
     :ivar statistics: If `show_stats=True` was specified in the request this
         field will contain information about the document payload.
     :vartype statistics: Optional[~azure.ai.textanalytics.TextDocumentStatistics]
+    :ivar detected_language: If 'language' was set to 'auto' for the document in the request this
+        field will contain a 2 letter ISO 639-1 representation of the language detected for this document.
+    :vartype detected_language: Optional[~azure.ai.textanalytics.DetectedLanguage]
     :ivar bool is_error: Boolean check for error item when iterating over list of
         results. Always False for an instance of a RecognizeCustomEntitiesResult.
     :ivar str kind: The text analysis kind - "CustomEntityRecognition".
+
+    .. versionadded:: 2022-10-01-preview
+        The *detected_language* property.
     """
 
     def __init__(self, **kwargs):
@@ -2254,15 +2322,18 @@ class RecognizeCustomEntitiesResult(DictMixin):
         self.entities = kwargs.get("entities", None)
         self.warnings = kwargs.get("warnings", [])
         self.statistics = kwargs.get("statistics", None)
+        self.detected_language = kwargs.get("detected_language", None)
         self.is_error: Literal[False] = False
         self.kind: Literal["CustomEntityRecognition"] = "CustomEntityRecognition"
 
     def __repr__(self):
-        return "RecognizeCustomEntitiesResult(id={}, entities={}, warnings={}, statistics={}, is_error={})".format(
+        return "RecognizeCustomEntitiesResult(id={}, entities={}, warnings={}, statistics={}, " \
+               "detected_language={}, is_error={})".format(
                 self.id,
                 repr(self.entities),
                 repr(self.warnings),
                 repr(self.statistics),
+                repr(self.detected_language),
                 self.is_error,
             )[
                 :1024
@@ -2285,6 +2356,9 @@ class RecognizeCustomEntitiesResult(DictMixin):
             statistics=TextDocumentStatistics._from_generated(  # pylint: disable=protected-access
                 result.statistics
             ),
+            detected_language=DetectedLanguage._from_generated(  # pylint: disable=protected-access
+                result.detected_language
+            ) if hasattr(result, "detected_language") and result.detected_language else None
         )
 
 
@@ -2359,9 +2433,15 @@ class ClassifyDocumentResult(DictMixin):
     :ivar statistics: If `show_stats=True` was specified in the request this
         field will contain information about the document payload.
     :vartype statistics: Optional[~azure.ai.textanalytics.TextDocumentStatistics]
+    :ivar detected_language: If 'language' was set to 'auto' for the document in the request this
+        field will contain a 2 letter ISO 639-1 representation of the language detected for this document.
+    :vartype detected_language: Optional[~azure.ai.textanalytics.DetectedLanguage]
     :ivar bool is_error: Boolean check for error item when iterating over list of
         results. Always False for an instance of a ClassifyDocumentResult.
     :ivar str kind: The text analysis kind - "CustomDocumentClassification".
+
+    .. versionadded:: 2022-10-01-preview
+        The *detected_language* property.
     """
 
     def __init__(
@@ -2372,16 +2452,18 @@ class ClassifyDocumentResult(DictMixin):
         self.classifications = kwargs.get('classifications', None)
         self.warnings = kwargs.get('warnings', [])
         self.statistics = kwargs.get('statistics', None)
+        self.detected_language = kwargs.get('detected_language', None)
         self.is_error: Literal[False] = False
         self.kind: Literal["CustomDocumentClassification"] = "CustomDocumentClassification"
 
     def __repr__(self):
-        return "ClassifyDocumentResult(id={}, classifications={}, warnings={}, statistics={}, " \
+        return "ClassifyDocumentResult(id={}, classifications={}, warnings={}, statistics={}, detected_language={} " \
                "is_error={})".format(
                 self.id,
                 repr(self.classifications),
                 repr(self.warnings),
                 repr(self.statistics),
+                repr(self.detected_language),
                 self.is_error,
             )[
                 :1024
@@ -2404,6 +2486,9 @@ class ClassifyDocumentResult(DictMixin):
             statistics=TextDocumentStatistics._from_generated(  # pylint: disable=protected-access
                 result.statistics
             ),
+            detected_language=DetectedLanguage._from_generated(  # pylint: disable=protected-access
+                result.detected_language
+            ) if hasattr(result, "detected_language") and result.detected_language else None
         )
 
 
@@ -2722,7 +2807,7 @@ class ExtractSummaryResult(DictMixin):
             ),
             detected_language=DetectedLanguage._from_generated(  # pylint: disable=protected-access
                 summary.detected_language
-            ) if summary.detected_language else None
+            )  if hasattr(summary, "detected_language") and summary.detected_language else None
         )
 
 
@@ -2822,7 +2907,7 @@ class AbstractSummaryResult(DictMixin):
             id=result.id,
             detected_language=DetectedLanguage._from_generated(  # pylint: disable=protected-access
                 result.detected_language
-            ) if result.detected_language else None,
+            ) if hasattr(result, "detected_language") and result.detected_language else None,
             warnings=[
                 TextAnalyticsWarning._from_generated(  # pylint: disable=protected-access
                     w