+++ sdk/textanalytics/azure-ai-textanalytics/azure/ai/textanalytics/_text_analytics_client.py
@@ -621,7 +621,7 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
     @validate_multiapi_args(
         version_method_added="v3.1",
         args_mapping={
-            "2022-10-01-preview": ["fhir_version", "document_type"],
+            "2022-10-01-preview": ["fhir_version", "document_type", "autodetect_default_language"],
             "2022-05-01": ["display_name"]
         }
     )
@@ -652,9 +652,12 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
         :keyword bool show_stats: If set to true, response will contain document level statistics.
         :keyword str language: The 2 letter ISO 639-1 representation of language for the
             entire batch. For example, use "en" for English; "es" for Spanish etc.
-            If not set, uses "en" for English as default. Per-document language will
-            take precedence over whole batch language. See https://aka.ms/talangs for
-            supported languages in Language API.
+            For automatic language detection, use "auto" (Only supported by API version
+            2022-10-01-preview and newer). If not set, uses "en" for English as default.
+            Per-document language will take precedence over whole batch language.
+            See https://aka.ms/talangs for supported languages in Language API.
+        :keyword autodetect_default_language: Default/fallback language to use for documents requesting
+            automatic language detection.
         :keyword str display_name: An optional display name to set for the requested analysis.
         :keyword str string_index_type: Specifies the method used to interpret string offsets.
             `UnicodeCodePoint`, the Python encoding, is the default. To override the Python default,
@@ -695,7 +698,7 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
         .. versionadded:: 2022-05-01
             The *display_name* keyword argument.
         .. versionadded:: 2022-10-01-preview
-            The *fhir_version* and *document_type* keyword arguments.
+            The *fhir_version*, *document_type*, and *autodetect_default_language* keyword arguments.
 
         .. admonition:: Example:
 
@@ -717,6 +720,7 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
         display_name = kwargs.pop("display_name", None)
         fhir_version = kwargs.pop("fhir_version", None)
         document_type = kwargs.pop("document_type", None)
+        autodetect_default_language = kwargs.pop("autodetect_default_language", None)
 
         if continuation_token:
             return cast(
@@ -759,6 +763,7 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
                         body=models.AnalyzeTextJobsInput(
                             analysis_input=docs,
                             display_name=display_name,
+                            default_language=autodetect_default_language,
                             tasks=[
                                 models.HealthcareLROTask(
                                     task_name="0",
@@ -1065,7 +1070,10 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
     @distributed_trace
     @validate_multiapi_args(
         version_method_added="v3.1",
-        custom_wrapper=check_for_unsupported_actions_types
+        custom_wrapper=check_for_unsupported_actions_types,
+        args_mapping={
+            "2022-10-01-preview": ["autodetect_default_language"],
+        }
     )
     def begin_analyze_actions(
         self,
@@ -1133,9 +1141,12 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
         :keyword str display_name: An optional display name to set for the requested analysis.
         :keyword str language: The 2 letter ISO 639-1 representation of language for the
             entire batch. For example, use "en" for English; "es" for Spanish etc.
-            If not set, uses "en" for English as default. Per-document language will
-            take precedence over whole batch language. See https://aka.ms/talangs for
-            supported languages in Language API.
+            For automatic language detection, use "auto" (Only supported by API version
+            2022-10-01-preview and newer). If not set, uses "en" for English as default.
+            Per-document language will take precedence over whole batch language.
+            See https://aka.ms/talangs for supported languages in Language API.
+        :keyword autodetect_default_language: Default/fallback language to use for documents requesting
+            automatic language detection.
         :keyword bool show_stats: If set to true, response will contain document level statistics.
         :keyword int polling_interval: Waiting time between two polls for LRO operations
             if no Retry-After header is present. Defaults to 5 seconds.
@@ -1172,6 +1183,7 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
         .. versionadded:: 2022-10-01-preview
             The *ExtractSummaryAction* and *AbstractSummaryAction* input options and the corresponding
             *ExtractSummaryResult* and *AbstractSummaryResult* result objects.
+            The *autodetect_default_language* keyword argument.
 
         .. admonition:: Example:
 
@@ -1191,6 +1203,7 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
         polling_interval = kwargs.pop("polling_interval", 5)
         language = language_arg if language_arg is not None else self._default_language
         bespoke = kwargs.pop("bespoke", False)
+        autodetect_default_language = kwargs.pop("autodetect_default_language", None)
 
         if continuation_token:
             return cast(
@@ -1246,6 +1259,7 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
                         body=models.AnalyzeTextJobsInput(
                             analysis_input=docs,
                             display_name=display_name,
+                            default_language=autodetect_default_language,
                             tasks=generated_tasks
                         ),
                         cls=response_cls,
@@ -1320,13 +1334,16 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
 
     @distributed_trace
     @validate_multiapi_args(
-        version_method_added="2022-05-01"
+        version_method_added="2022-05-01",
+        args_mapping={
+            "2022-10-01-preview": ["autodetect_default_language"],
+        }
     )
     def begin_recognize_custom_entities(
         self,
         documents: Union[List[str], List[TextDocumentInput], List[Dict[str, str]]],
-        project_name,
-        deployment_name,
+        project_name: str,
+        deployment_name: str,
         **kwargs: Any,
     ) -> TextAnalysisLROPoller[ItemPaged[Union[RecognizeCustomEntitiesResult, DocumentError]]]:
         """Start a long-running custom named entity recognition operation.
@@ -1345,9 +1362,12 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
         :param str deployment_name: This field indicates the deployment name for the model.
         :keyword str language: The 2 letter ISO 639-1 representation of language for the
             entire batch. For example, use "en" for English; "es" for Spanish etc.
-            If not set, uses "en" for English as default. Per-document language will
-            take precedence over whole batch language. See https://aka.ms/talangs for
-            supported languages in Language API.
+            For automatic language detection, use "auto" (Only supported by API version
+            2022-10-01-preview and newer). If not set, uses "en" for English as default.
+            Per-document language will take precedence over whole batch language.
+            See https://aka.ms/talangs for supported languages in Language API.
+        :keyword autodetect_default_language: Default/fallback language to use for documents requesting
+            automatic language detection.
         :keyword bool show_stats: If set to true, response will contain document level statistics.
         :keyword bool disable_service_logs: If set to true, you opt-out of having your text input
             logged on the service side for troubleshooting. By default, the Language service logs your
@@ -1379,6 +1399,8 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
 
         .. versionadded:: 2022-05-01
             The *begin_recognize_custom_entities* client method.
+        .. versionadded:: 2022-10-01-preview
+            The *autodetect_default_language* keyword argument.
 
         .. admonition:: Example:
 
@@ -1438,13 +1460,16 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
 
     @distributed_trace
     @validate_multiapi_args(
-        version_method_added="2022-05-01"
+        version_method_added="2022-05-01",
+        args_mapping={
+            "2022-10-01-preview": ["autodetect_default_language"],
+        }
     )
     def begin_single_label_classify(
         self,
         documents: Union[List[str], List[TextDocumentInput], List[Dict[str, str]]],
-        project_name,
-        deployment_name,
+        project_name: str,
+        deployment_name: str,
         **kwargs: Any,
     ) -> TextAnalysisLROPoller[ItemPaged[Union[ClassifyDocumentResult, DocumentError]]]:
         """Start a long-running custom single label classification operation.
@@ -1463,9 +1488,12 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
         :param str deployment_name: This field indicates the deployment name for the model.
         :keyword str language: The 2 letter ISO 639-1 representation of language for the
             entire batch. For example, use "en" for English; "es" for Spanish etc.
-            If not set, uses "en" for English as default. Per-document language will
-            take precedence over whole batch language. See https://aka.ms/talangs for
-            supported languages in Language API.
+            For automatic language detection, use "auto" (Only supported by API version
+            2022-10-01-preview and newer). If not set, uses "en" for English as default.
+            Per-document language will take precedence over whole batch language.
+            See https://aka.ms/talangs for supported languages in Language API.
+        :keyword autodetect_default_language: Default/fallback language to use for documents requesting
+            automatic language detection.
         :keyword bool show_stats: If set to true, response will contain document level statistics.
         :keyword bool disable_service_logs: If set to true, you opt-out of having your text input
             logged on the service side for troubleshooting. By default, the Language service logs your
@@ -1493,6 +1521,8 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
 
         .. versionadded:: 2022-05-01
             The *begin_single_label_classify* client method.
+        .. versionadded:: 2022-10-01-preview
+            The *autodetect_default_language* keyword argument.
 
         .. admonition:: Example:
 
@@ -1550,13 +1580,16 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
 
     @distributed_trace
     @validate_multiapi_args(
-        version_method_added="2022-05-01"
+        version_method_added="2022-05-01",
+        args_mapping={
+            "2022-10-01-preview": ["autodetect_default_language"],
+        }
     )
     def begin_multi_label_classify(
         self,
         documents: Union[List[str], List[TextDocumentInput], List[Dict[str, str]]],
-        project_name,
-        deployment_name,
+        project_name: str,
+        deployment_name: str,
         **kwargs: Any,
     ) -> TextAnalysisLROPoller[ItemPaged[Union[ClassifyDocumentResult, DocumentError]]]:
         """Start a long-running custom multi label classification operation.
@@ -1575,9 +1608,12 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
         :param str deployment_name: This field indicates the deployment name for the model.
         :keyword str language: The 2 letter ISO 639-1 representation of language for the
             entire batch. For example, use "en" for English; "es" for Spanish etc.
-            If not set, uses "en" for English as default. Per-document language will
-            take precedence over whole batch language. See https://aka.ms/talangs for
-            supported languages in Language API.
+            For automatic language detection, use "auto" (Only supported by API version
+            2022-10-01-preview and newer). If not set, uses "en" for English as default.
+            Per-document language will take precedence over whole batch language.
+            See https://aka.ms/talangs for supported languages in Language API.
+        :keyword autodetect_default_language: Default/fallback language to use for documents requesting
+            automatic language detection.
         :keyword bool show_stats: If set to true, response will contain document level statistics.
         :keyword bool disable_service_logs: If set to true, you opt-out of having your text input
             logged on the service side for troubleshooting. By default, the Language service logs your
@@ -1605,6 +1641,8 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
 
         .. versionadded:: 2022-05-01
             The *begin_multi_label_classify* client method.
+        .. versionadded:: 2022-10-01-preview
+            The *autodetect_default_language* keyword argument.
 
         .. admonition:: Example: