+++ sdk/textanalytics/azure-ai-textanalytics/azure/ai/textanalytics/aio/_text_analytics_client_async.py
@@ -838,7 +838,7 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
     @validate_multiapi_args(
         version_method_added="v3.1",
         args_mapping={
-            "2022-10-01-preview": ["fhir_version", "document_type"],
+            "2022-10-01-preview": ["fhir_version", "document_type", "autodetect_default_language"],
             "2022-05-01": ["display_name"]
         }
     )
@@ -871,9 +871,12 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
         :keyword bool show_stats: If set to true, response will contain document level statistics.
         :keyword str language: The 2 letter ISO 639-1 representation of language for the
             entire batch. For example, use "en" for English; "es" for Spanish etc.
-            If not set, uses "en" for English as default. Per-document language will
-            take precedence over whole batch language. See https://aka.ms/talangs for
-            supported languages in Text Analytics API.
+            For automatic language detection, use "auto" (Only supported by API version
+            2022-10-01-preview and newer). If not set, uses "en" for English as default.
+            Per-document language will take precedence over whole batch language.
+            See https://aka.ms/talangs for supported languages in Language API.
+        :keyword autodetect_default_language: Default/fallback language to use for documents requesting
+            automatic language detection.
         :keyword str display_name: An optional display name to set for the requested analysis.
         :keyword str string_index_type: Specifies the method used to interpret string offsets.
             Can be one of 'UnicodeCodePoint' (default), 'Utf16CodeUnit', or 'TextElement_v8'.
@@ -913,7 +916,7 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
         .. versionadded:: 2022-05-01
             The *display_name* keyword argument.
         .. versionadded:: 2022-10-01-preview
-            The *fhir_version* and *document_type* keyword arguments.
+            The *fhir_version*, *document_type*, and *autodetect_default_language* keyword arguments.
 
         .. admonition:: Example:
 
@@ -935,6 +938,7 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
         display_name = kwargs.pop("display_name", None)
         fhir_version = kwargs.pop("fhir_version", None)
         document_type = kwargs.pop("document_type", None)
+        autodetect_default_language = kwargs.pop("autodetect_default_language", None)
 
         if continuation_token:
             return cast(
@@ -977,6 +981,7 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
                         body=models.AnalyzeTextJobsInput(
                             analysis_input=docs,
                             display_name=display_name,
+                            default_language=autodetect_default_language,
                             tasks=[
                                 models.HealthcareLROTask(
                                     task_name="0",
@@ -1060,7 +1065,10 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
     @distributed_trace_async
     @validate_multiapi_args(
         version_method_added="v3.1",
-        custom_wrapper=check_for_unsupported_actions_types
+        custom_wrapper=check_for_unsupported_actions_types,
+        args_mapping={
+            "2022-10-01-preview": ["autodetect_default_language"],
+        }
     )
     async def begin_analyze_actions(
         self,
@@ -1128,9 +1136,12 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
         :keyword str display_name: An optional display name to set for the requested analysis.
         :keyword str language: The 2 letter ISO 639-1 representation of language for the
             entire batch. For example, use "en" for English; "es" for Spanish etc.
-            If not set, uses "en" for English as default. Per-document language will
-            take precedence over whole batch language. See https://aka.ms/talangs for
-            supported languages in Language API.
+            For automatic language detection, use "auto" (Only supported by API version
+            2022-10-01-preview and newer). If not set, uses "en" for English as default.
+            Per-document language will take precedence over whole batch language.
+            See https://aka.ms/talangs for supported languages in Language API.
+        :keyword autodetect_default_language: Default/fallback language to use for documents requesting
+            automatic language detection.
         :keyword bool show_stats: If set to true, response will contain document level statistics.
         :keyword int polling_interval: Waiting time between two polls for LRO operations
             if no Retry-After header is present. Defaults to 5 seconds.
@@ -1167,6 +1178,7 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
         .. versionadded:: 2022-10-01-preview
             The *ExtractSummaryAction* and *AbstractSummaryAction* input options and the corresponding
             *ExtractSummaryResult* and *AbstractSummaryResult* result objects.
+            The *autodetect_default_language* keyword argument.
 
         .. admonition:: Example:
 
@@ -1187,6 +1199,7 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
         polling_interval = kwargs.pop("polling_interval", 5)
         continuation_token = kwargs.pop("continuation_token", None)
         bespoke = kwargs.pop("bespoke", False)
+        autodetect_default_language = kwargs.pop("autodetect_default_language", None)
 
         if continuation_token:
             return cast(
@@ -1243,6 +1256,7 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
                         body=models.AnalyzeTextJobsInput(
                             analysis_input=docs,
                             display_name=display_name,
+                            default_language=autodetect_default_language,
                             tasks=generated_tasks
                         ),
                         cls=response_cls,
@@ -1317,7 +1331,10 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
 
     @distributed_trace_async
     @validate_multiapi_args(
-        version_method_added="2022-05-01"
+        version_method_added="2022-05-01",
+        args_mapping={
+            "2022-10-01-preview": ["autodetect_default_language"],
+        }
     )
     async def begin_recognize_custom_entities(
         self,
@@ -1342,9 +1359,12 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
         :param str deployment_name: This field indicates the deployment name for the model.
         :keyword str language: The 2 letter ISO 639-1 representation of language for the
             entire batch. For example, use "en" for English; "es" for Spanish etc.
-            If not set, uses "en" for English as default. Per-document language will
-            take precedence over whole batch language. See https://aka.ms/talangs for
-            supported languages in Language API.
+            For automatic language detection, use "auto" (Only supported by API version
+            2022-10-01-preview and newer). If not set, uses "en" for English as default.
+            Per-document language will take precedence over whole batch language.
+            See https://aka.ms/talangs for supported languages in Language API.
+        :keyword autodetect_default_language: Default/fallback language to use for documents requesting
+            automatic language detection.
         :keyword bool show_stats: If set to true, response will contain document level statistics.
         :keyword bool disable_service_logs: If set to true, you opt-out of having your text input
             logged on the service side for troubleshooting. By default, the Language service logs your
@@ -1376,6 +1396,8 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
 
         .. versionadded:: 2022-05-01
             The *begin_recognize_custom_entities* client method.
+        .. versionadded:: 2022-10-01-preview
+            The *autodetect_default_language* keyword argument.
 
         .. admonition:: Example:
 
@@ -1435,7 +1457,10 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
 
     @distributed_trace_async
     @validate_multiapi_args(
-        version_method_added="2022-05-01"
+        version_method_added="2022-05-01",
+        args_mapping={
+            "2022-10-01-preview": ["autodetect_default_language"],
+        }
     )
     async def begin_single_label_classify(
         self,
@@ -1460,9 +1485,12 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
         :param str deployment_name: This field indicates the deployment name for the model.
         :keyword str language: The 2 letter ISO 639-1 representation of language for the
             entire batch. For example, use "en" for English; "es" for Spanish etc.
-            If not set, uses "en" for English as default. Per-document language will
-            take precedence over whole batch language. See https://aka.ms/talangs for
-            supported languages in Language API.
+            For automatic language detection, use "auto" (Only supported by API version
+            2022-10-01-preview and newer). If not set, uses "en" for English as default.
+            Per-document language will take precedence over whole batch language.
+            See https://aka.ms/talangs for supported languages in Language API.
+        :keyword autodetect_default_language: Default/fallback language to use for documents requesting
+            automatic language detection.
         :keyword bool show_stats: If set to true, response will contain document level statistics.
         :keyword bool disable_service_logs: If set to true, you opt-out of having your text input
             logged on the service side for troubleshooting. By default, the Language service logs your
@@ -1490,6 +1518,8 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
 
         .. versionadded:: 2022-05-01
             The *begin_single_label_classify* client method.
+        .. versionadded:: 2022-10-01-preview
+            The *autodetect_default_language* keyword argument.
 
         .. admonition:: Example:
 
@@ -1547,7 +1577,10 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
 
     @distributed_trace_async
     @validate_multiapi_args(
-        version_method_added="2022-05-01"
+        version_method_added="2022-05-01",
+        args_mapping={
+            "2022-10-01-preview": ["autodetect_default_language"],
+        }
     )
     async def begin_multi_label_classify(
         self,
@@ -1572,9 +1605,12 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
         :param str deployment_name: This field indicates the deployment name for the model.
         :keyword str language: The 2 letter ISO 639-1 representation of language for the
             entire batch. For example, use "en" for English; "es" for Spanish etc.
-            If not set, uses "en" for English as default. Per-document language will
-            take precedence over whole batch language. See https://aka.ms/talangs for
-            supported languages in Language API.
+            For automatic language detection, use "auto" (Only supported by API version
+            2022-10-01-preview and newer). If not set, uses "en" for English as default.
+            Per-document language will take precedence over whole batch language.
+            See https://aka.ms/talangs for supported languages in Language API.
+        :keyword autodetect_default_language: Default/fallback language to use for documents requesting
+            automatic language detection.
         :keyword bool show_stats: If set to true, response will contain document level statistics.
         :keyword bool disable_service_logs: If set to true, you opt-out of having your text input
             logged on the service side for troubleshooting. By default, the Language service logs your
@@ -1602,6 +1638,8 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
 
         .. versionadded:: 2022-05-01
             The *begin_multi_label_classify* client method.
+        .. versionadded:: 2022-10-01-preview
+            The *autodetect_default_language* keyword argument.
 
         .. admonition:: Example: